v1

+++ src/main/java/com/examly/springapp/repository/InstructorRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.examly.springapp.model.Instructor;
 
 @Repository
-public interface InstructorRepo extends JpaRepository<Instructor,Integer>{
+public interface InstructorRepo extends JpaRepository<Instructor,Long>{
     
     List<Instructor> findBySpecialization(String specialization);
